В main уменьшил данные и сразу вписал Юзеров в usersCompare.
